made some changes

+++ main.py
@@ -2,7 +2,7 @@ print('for basic  math, choose option 1')
 print('for trig calculations, choose option 2')
 print('for unit conversion, choose option 3')
 print('to exit   choose option 4')
-opt = int(input())
+opt = float(input())
 
 if opt==1:
  import basic_math
