+++ trig_calculator.py
@@ -2,21 +2,21 @@ import math
 
 def calculate_sin():
     print("Input a degree to convert to sine value")
-    degree = int(input())
+    degree = float(input())
     result = degree * (22/7)/180
     res2 = math.sin(result)
     print(res2)
 
 def calculate_cos():
     print("Input a degree to convert to cosine value")
-    degree = int(input())
+    degree = float(input())
     result = degree * (22/7)/180
     res2 = math.cos(result)
     print(res2)
 
 def calculate_tan():
     print("Input a degree to convert to tangent value")
-    degree = int(input())
+    degree = float(input())
     result = degree * (22/7)/180
     res2 = math.tan(result)
     print(res2)
@@ -25,7 +25,7 @@ print("Select option 1 to find sine value")
 print("Select option 2 to find cosine value")
 print("Select option 3 to find tangent value")
 
-opt = int(input("Enter your choice: "))
+opt = float(input("Enter your choice: "))
 
 if opt == 1:
     calculate_sin()
